Google authetication added to App's state.
Things got real. Wheather a user is loggged in or not is an app-wide information.
So app.js got rehauled into class component to hold the state.

Additionally two lifecycle methods are added - Component did Mount connects to Google firebase auth
and sets up a way to close the constantly open connection.

That job falls to componentWillUnmount, so no memory leak is left in form of connection go firebase while app is already gone.